+++ packages/types/custom/core.ts
@@ -559,12 +559,14 @@ export interface MapConfig {
   namedProjections: Array<[string, string]>
   /** Mapped resolution to zoomLevel */
   options?: PolarMapOptions[]
+  renderFaToLightDom?: boolean
   /** Extent in which the map can be viewed in; coordinates are written in the set projection of the map set through this config. */
   extent?: number[]
   language?: InitialLanguage
   locales?: LanguageOption[]
   /** Initial center coordinate for the mapView */
   startCenter?: number[]
+  stylePath?: string
   /** Initial resolution the map should be rendered with */
   startResolution?: number
   vuetify?: UserVuetifyPreset